isLucky
First I needed to find a way to split up the input number so that we can search through and operate on individual numbers. I did this by turning n into a character string, using strsplit() on that and then turning it back into a numeric vector. I then needed to find a way to split the numeric vector in half such that we can check the sum of the first half of elements against the sum of the second half.  I tried a couple mistaken methods, but finally settled on using the split function, which works.